updated new input

+++ pages/input.js
@@ -102,11 +102,11 @@ export default function Input() {
 			// garbage_count: parseInt(localStorage.garbageCount),
 			// compost_count: parseInt(localStorage.compostCount),
 			// recycling_count: parseInt(localStorage.recyclingCount),
-			item_1_count: parseInt(localStorage.item_1_count),
-			item_2_count: parseInt(localStorage.item_2_count),
-			item_3_count: parseInt(localStorage.item_3_count),
-			item_4_count: parseInt(localStorage.item_4_count),
-			item_5_count: parseInt(localStorage.item_5_count),
+			item_1_count: parseFloat(localStorage.item_1_count),
+			item_2_count: parseFloat(localStorage.item_2_count),
+			item_3_count: parseFloat(localStorage.item_3_count),
+			item_4_count: parseFloat(localStorage.item_4_count),
+			item_5_count: parseFloat(localStorage.item_5_count),
 		};
 
 		/* No longer making a post request this way */
